create common TestCase for subscriber test & refactor
