+++ tests/Integration/inc/Engine/CDN/Subscriber/rewriteCssProperties.php
@@ -2,8 +2,6 @@
 
 namespace WP_Rocket\Tests\Integration\inc\Engine\CDN\Subscriber;
 
-use WPMedia\PHPUnit\Integration\TestCase;
-
 /**
  * @covers \WP_Rocket\Engine\CDN\Subscriber::rewrite_css_properties
  * @uses   \WP_Rocket\Engine\CDN\CDN::rewrite_css_properties
@@ -11,9 +9,6 @@ use WPMedia\PHPUnit\Integration\TestCase;
  * @group  CDN
  */
 class Test_RewriteCssProperties extends TestCase {
-	private $cnames;
-	private $cdn_zone;
-
 	public function setUp() {
 		parent::setUp();
 
@@ -30,9 +25,6 @@ class Test_RewriteCssProperties extends TestCase {
 	}
 
 	public function tearDown() {
-		remove_filter( 'pre_get_rocket_option_cdn', [ $this, 'return_true' ] );
-		remove_filter( 'pre_get_rocket_option_cdn_cnames', [ $this, 'setCnames' ] );
-		remove_filter( 'pre_get_rocket_option_cdn_zone', [ $this, 'setCDNZone' ] );
 		remove_filter( 'do_rocket_cdn_css_properties', [ $this, 'return_false' ] );
 
 		parent::tearDown();
@@ -62,13 +54,5 @@ class Test_RewriteCssProperties extends TestCase {
 
 	public function providerTestData() {
 		return $this->getTestData( __DIR__, 'rewriteCssProperties' );
-	}
-
-	public function setCnames() {
-		return $this->cnames;
-	}
-
-	public function setCDNZone() {
-		return $this->cdn_zone;
 	}
 }
